requirements.txt
loguru
omegaconf
tqdm
smplx
soundfile
lmdb
librosa
PyYAML
wandb
tensorboard
pynvml
matplotlib
configargparse
pandas
opencv-python
fasttest
IPython
h5py
einops
textgrid
termcolor
transformers
gdown
fasttext
pyrender
trimesh
imageio
moviepy
huggingface_hub
diffusers==0.24.0
numpy==1.24.2